Upgrade clipgen to Python 3 using 2to3.
2to3 ran successfully, however two minor adjustments needed to be made manually.

The minor adjustments:

First; commented out the encoding fix at the start of the program, lines 9-10:
> # Dirty unicode encoding fix
> #importlib.reload(sys)
> #sys.setdefaultencoding('utf8')

Second; the study number detection on line 630 was concatting str with list objects. In Python 3, this is no longer allowed. Added a dirty fix of selecting the first index of the list (which was fortunately a str) and sent that back instead:
> ... list(filter(str.isdigit, timesList[i]['study']))[0] ...